+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to
 
 - 🎨(frontend) better conversion editor to pdf #151
 - ✨(frontend) Versioning #147
+- ✨Export docx (word) #161
 
 ## Fixed