+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to
 - 🎨(frontend) better conversion editor to pdf #151
 - ✨(frontend) Versioning #147
 - ✨Export docx (word) #161
+- 🌐 Internationalize invitation email #167
 
 ## Fixed